+++ CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
+## [6.0.0] - 2024-09-21
 ### Changed
+- Replaced Apache CloseableHttpClient with configurable azure-core client.
+- [BREAKING] HttpClientFactory now accepts clients implementing azure-core HttpClient.
+- [BREAKING] HttpClientProperties and HttpClientPropertiesBuilder now use azure-core ProxyOptions.
+- Data client now wraps internal HTTP client.
+- Moved HTTP request tracing logic into a builder class.
+- Moved HTTP request building logic into a builder class.
 - [BREAKING] Redirects are disabled by default. Use ClientRequestProperties "client_max_redirect_count" option 
     to enable. Default changed to 0.
+
 ## [5.2.0] - 2024-08-27
 ### Fixed
 - Used Msal user prompt old code which is deprecated in the new version coming from last bom update resulted in method not found exception.